+++ src/components/SVGIcon.vue
@@ -1,16 +1,18 @@
 <template>
-    <component :is="dynamicComponent" />
+  <component :is="dynamicComponent" />
 </template>
 
 <script setup>
-import { defineAsyncComponent } from 'vue';
+import { defineAsyncComponent } from "vue";
 
 const props = defineProps({
-    name: {
-        type: String,
-        required: true,
-    },
-})
+  name: {
+    type: String,
+    required: true,
+  },
+});
 
-const dynamicComponent = defineAsyncComponent(() => import(`../assets/svg/${props.name}.svg?component`));
+const dynamicComponent = defineAsyncComponent(
+  () => import(`../assets/svg/${props.name}.svg?component`)
+);
 </script>