+++ src/main.js
@@ -1,11 +1,11 @@
-import { createApp } from 'vue'
-import '@/assets/css/main-style.css'
-import '@/assets/css/card.css'
-import '@/assets/css/dark-mode.css'
-import '@/assets/css/links.css'
-import '@/assets/css/modal.css'
-import '@/assets/css/button.css'
-import App from './App.vue'
+import { createApp } from "vue";
+import "@/assets/css/main-style.css";
+import "@/assets/css/card.css";
+import "@/assets/css/dark-mode.css";
+import "@/assets/css/links.css";
+import "@/assets/css/modal.css";
+import "@/assets/css/button.css";
+import "@/assets/css/tooltip.css";
+import App from "./App.vue";
 
-createApp(App)
-  .mount('#app')
+createApp(App).mount("#app");